[REFACTOR] : ♻️ 다같이 / datatest -> DataManager 로 코드 리팩토링
