+++ GYMGLE/Pages/Admin/AdminNoticePage/ViewController/AdminNoticeViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 final class AdminNoticeViewController: UIViewController {
 
     // MARK: - dummyData
-    private let dummyDataManager = DataTest.shared
+    private let dummyDataManager = DataManager.shared
     var gymInfo: GymInfo?
 
     private let adminNoticeView = AdminNoticeView()
@@ -74,14 +74,14 @@ extension AdminNoticeViewController {
 
 extension AdminNoticeViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return dummyDataManager.test.noticeList.count
+        return dummyDataManager.gymInfo.noticeList.count
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: AdminNoticeTableViewCell.identifier, for: indexPath) as! AdminNoticeTableViewCell
-        var date = dateToString(date: dummyDataManager.test.noticeList[indexPath.row].date)
+        var date = dateToString(date: dummyDataManager.gymInfo.noticeList[indexPath.row].date)
         cell.nameLabel.text = gymInfo?.gymName
-        cell.contentLabel.text = dummyDataManager.test.noticeList[indexPath.row].content
+        cell.contentLabel.text = dummyDataManager.gymInfo.noticeList[indexPath.row].content
         cell.dateLabel.text = date
         
         cell.selectionStyle = .none
@@ -94,7 +94,7 @@ extension AdminNoticeViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
        
         let adminNoticeDetailVC = AdminNoticeDetailViewController()
-        adminNoticeDetailVC.noticeInfo = dummyDataManager.test.noticeList[indexPath.row]
+        adminNoticeDetailVC.noticeInfo = dummyDataManager.gymInfo.noticeList[indexPath.row]
         navigationController?.pushViewController(adminNoticeDetailVC, animated: true)
     }
 }
